Split C header
